# travel-api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Travel Api",
    "contact": {}
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users": {
      "get": {
        "description": "- URL\n    - /users\n- Method\n    - GET\n- Headers\n    - x-access-token : Bearer\n- Response\n    \n\n<img src=\"https://content.pstmn.io/f9657763-06dd-4f04-a08c-815c5adfe648/aW1hZ2UucG5n\" width=\"698\" height=\"561\">",
        "summary": "Get all users",
        "tags": [
          "users"
        ],
        "operationId": "Getallusers",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "description": "- URL\n    - /users/\n- Method\n    - GET\n- Headers\n    - x-access-token : Bearer\n- Response\n    \n\n<img src=\"https://content.pstmn.io/1343952d-2edf-40d8-85ad-193fb9c9493a/aW1hZ2UucG5n\" width=\"681\" height=\"391\">",
        "summary": "Get user by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "users"
        ],
        "operationId": "GetuserbyID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "description": "- URL\n    - /users/\n- Method\n    - PUT\n- Headers\n    - x-access-token : Bearer\n- Response\n    \n\n<img src=\"https://content.pstmn.io/1343952d-2edf-40d8-85ad-193fb9c9493a/aW1hZ2UucG5n\" width=\"681\" height=\"391\">",
        "summary": "Get user by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "users"
        ],
        "operationId": "PutuserbyID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "description": "",
            "schema": {
              "$ref": "#/definitions/EditUserRequest"
            }
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/json"
            ],
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "- URL\n    - /users/\n- Method\n    - DELETE\n- Headers\n    - x-access-token : Bearer\n- Response\n    \n\n<img src=\"https://content.pstmn.io/eb5c6f81-83d7-46c1-96ee-66f19c81e68e/aW1hZ2UucG5n\" alt=\"\" height=\"176\" width=\"704\">",
        "summary": "Delete user by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "users"
        ],
        "operationId": "DeleteuserbyID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "EditUserRequest": {
      "title": "EditRequest",
      "example": {
        "name": "update_name",
        "email": "update_email",
        "old_password": "old_password",
        "new_password": "new_password",
        "telephone": "update_telephone",
        "address": "update_address",
        "role": "update_role"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "telephone": {
          "type": "string"
        },
        "old_password": {
          "type": "string"
        },
        "new_password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
...
  },
...
}